* marshal.c (r_object): better allocation type check for   TYPE_UCLASS. usage of allocation framework is disabled for now.
* variable.c (rb_class_path): Module may have subclass.

* string.c (rb_str_update): should maintain original negative
  offset.

* string.c (rb_str_subpat_set): ditto

* string.c (rb_str_aset): ditto.

* re.c (rb_reg_nth_match): should check negative nth.

* re.c (rb_reg_nth_defined): ditto.


git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@1764 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ regex.c
@@ -3412,7 +3412,7 @@ re_search(bufp, string, size, startpos, range, regs)
 									\
     *stackp++ = pattern_place;                                          \
     *stackp++ = string_place;                                           \
-    *stackp++ = (unsigned char*)options; /* current option status */	\
+    *stackp++ = (unsigned char*)(long)options; /* current option status */	\
     *stackp++ = (unsigned char*)0; /* non-greedy flag */		\
   } while(0)
 
@@ -4293,7 +4293,7 @@ re_match(bufp, string_arg, size, pos, regs)
 	goto fail;
       }
       stackp--;		/* discard greedy flag */
-      options = (int)*--stackp;
+      options = (long)*--stackp;
       d = *--stackp;
       p = *--stackp;
       /* Restore register info.  */
